pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarket-whale-bot"
version = "0.1.0"
description = "Polymarket whale-watching copy trading bot"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "aiohttp>=3.9.3,<4.0",
    "sqlalchemy[asyncio]>=2.0.25,<3.0",
    "asyncpg>=0.29.0,<1.0",
    "redis[hiredis]>=5.0.1,<6.0",
    "apscheduler>=3.10.4,<4.0",
    "pydantic>=2.5.3,<3.0",
    "pydantic-settings>=2.1.0,<3.0",
    "python-telegram-bot>=20.7,<21.0",
    "py-clob-client>=0.18.0",
    "structlog>=24.1.0,<25.0",
    "web3>=6.13.0,<7.0",
    "eth-account>=0.10.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
    "statistics; python_version < '3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-mock>=3.12.0",
    "black>=24.1.1",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "types-redis>=4.6.0.20240106",
]

[project.scripts]
polymarket-bot = "main:run"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501", "B008", "UP007"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
